Factor NetworkGraph logic into utilities method (getMaxUnit AndFormatNetwork)
