+++ packages/manager/src/features/Longview/LongviewDetail/DetailTabs/Network/NetworkGraphs.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { withTheme, WithTheme } from 'src/components/core/styles';
 import LongviewLineGraph from 'src/components/LongviewLineGraph';
-import { generateUnits } from 'src/features/Longview/LongviewLanding/Gauges/Network';
-import { statMax } from 'src/features/Longview/shared/utilities';
+import { getMaxUnitAndFormatNetwork } from 'src/features/Longview/shared/utilities';
 import { LongviewNetworkInterface } from '../../../request.types';
 import { convertData } from '../../../shared/formatters';
 import GraphCard from '../../GraphCard';
@@ -30,7 +29,6 @@ export const NetworkGraphs: React.FC<CombinedProps> = props => {
     theme,
     timezone
   } = props;
-  // const classes = useStyles();
 
   const _convertData = React.useCallback(convertData, [
     networkData,
@@ -46,26 +44,10 @@ export const NetworkGraphs: React.FC<CombinedProps> = props => {
         const [name, interfaceData] = thisInterface;
         const { rx_bytes, tx_bytes } = interfaceData;
 
-        // Determine the unit based on the largest value.
-        const max = Math.max(statMax(rx_bytes), statMax(tx_bytes));
-        const maxUnit = generateUnits(max).unit;
-
-        const formatNetwork = (valueInBytes: number | null) => {
-          if (valueInBytes === null) {
-            return valueInBytes;
-          }
-          const valueInBits = valueInBytes * 8;
-
-          if (maxUnit === 'Mb') {
-            // If the unit we're using for the graph is Mb, return the output in Mb.
-            const valueInMegabits = valueInBits / 1024 / 1024;
-            return Math.round(valueInMegabits * 100) / 100;
-          } else {
-            // If the unit we're using for the graph is Kb, return the output in Kb.
-            const valueInKilobits = valueInBits / 1024;
-            return Math.round(valueInKilobits * 100) / 100;
-          }
-        };
+        const { maxUnit, formatNetwork } = getMaxUnitAndFormatNetwork(
+          rx_bytes,
+          tx_bytes
+        );
 
         return (
           <GraphCard title={name} key={`network-interface-card-${idx}`}>
